Removed obsolete SUNCOSB array
The SUNCOSB array was used in physproc.f to compute the
photolysis at the next chemistry timestep for the purpose
of reordering grid boxes by stiffness.  However, this
reordering proved to be computationally inefficient, and
was never implemented.

With the computation of the photolysis at the midpt of
the chemistry timestep, SUNCOSB is now obsolete.  Delete
it from all relevant locations (chemdr.f, main.f, and
physproc.f)

Also add .gitignore files to every subdirectory to tell Git
to ignore files having certain extensions.  These settings
will be saved as part of the repository and will be applied
for everyone who clones the repository.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>