Minor fix for AOD calculation
Basically the extinction efficiency for each RH bin is
currently being calculated as the water volume-weighted
average of each bin's RH and dry value, when the value
corresponding to each bin's RH should simply be directly used.
The implications should not be large (Colette estimated
between 4-8%), but this ought to be corrected.

As a result, the variable FWET is no longer used and no
longer needs to be declared in RDAER.

Signed-off-by: Aaron van Donkelaar <Aaron.van.Donkelaar@dal.ca>
Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>