Add modifications for MERRA in Headers and GeosUtil dirs
Add #if defined( MERRA ) to various locations in the
define.h and CMN_SIZE header files, as well as to the
modules in GeosUtil.

We shall treat MERRA exactly like GEOS-5, since the vertical
grids are identical.  Therefore in many instances all that
was necessary was to change #ifdef strings such as

   #if defined( GEOS_5 )

to

   #if defined( GEOS_5 ) || defined( MERRA )


Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>