Add modifications for MERRA in GeosCore directory
Add #if defined( MERRA ) to various files in the GeosCore
directory.  For now we shall treat MERRA exactly like
GEOS-5.

The met field modules (e.g. dao_mod.f, *_read_mod.f, etc.)
were not updated, as this will require special handling.
They will be updated in the next commit.

The MERRA SNOMAS field needs to be validated against that
of GEOS-5.  There seems to be a units difference in the
MERRA files.

Also added ProTeX documentation headers to many of the
routines that were modified.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>