Now treat MERRA SNOMAS field the same as for GEOS-5
The MERRA SNOMAS (snow depth) field did not contain snow coverage
over the Antarctic and Greenland.  As is stated in the relevant
documentation:

  From the GEOS-5 File Specification Document:

    SNOMAS: The mass of snow in per unit of land area in
    meters of liquid-water-equivalent depth (i.e., 10^3 kg/m2).
    In grid boxes with no land (FRLAND+FRLANDICE=0) it is set
    to _FillValue (= 1e15).  Where FRLANDICE>0.9 it is
    arbitrarily set to 4 meters. Over other land areas it
    represents an average over the non-glaciated part.

  From the MERRA File Specification Document:
                                                                                 .
    SNOMAS: The mass of snow per unit of ice-free land area
    (FRLAND), in kg/m2. In grid boxes with no land it is set
    to _FillValue (=1e15).  Over other land areas it represents
    an average over the nonglaciated part.

To correct for this, the MERRA data processing code now applies
a default value of 4000 mm H2O where FRLANDICE > 0.9.  This will
bring the MERRA SNOMAS field back into agreement with what was
done in GEOS-5.

Therefore, in GEOS-Chem, we can treat the MERRA SNOMAS field as
in the same way as we do for GEOS-5.  Thus we can simply extend
the #if block statements from:

   #if defined( GEOS_5 )

to:

   #if defined( GEOS_5 ) || defined( MERRA )

in the various GEOS-Chem source code routins that refer to SNOMAS.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>