Temporarily disable Qiaoqiao Wang's washout/rainout modifications
Chris Holmes wrote:

   Qiaoqiao Wang implemented several changes to wet scavenging
   that I know of:

   1. Scavenging by snow has different collection efficiency
      to rain

   2. Allow both washout and rainout when QQ > 0 and
      FTOP > F_PRIME.  Previously only rainout occurred when
      QQ > 0.  Qiaoqiao reasoned that if QQ > 0 is very small
      and there is a lot of rain from above, then most of
      the box should experience washout and only some of the
      box should experience rainout.

   3. Specific improvements for BC and OC.

   I reorganized the code (in routine WETDEP) so that
   Qiaoqiao's changes for Item 2 could be enabled or disabled
   by commenting just a few lines.

To recap, Item #1 (scavenging by snow) and Item #3 (BC/OC)
were not implemented in v8-03-02.

Item #2 (washout/rainout improvements) was implemented by Chris
Holmes and activated in routine WETDEP.  However, by request of
Daniel Jacob, we have reverted to the previous behavior from
v8-03-01.  We are waiting on Qiaoqiao's paper to be submitted
and approved before "officially" adding Item #2 to the code.

Bob Yantosca added extra comments to clarify the places in
routine WETDEP where code has to be commented/uncommented
in order to invoke (or suppress) the Item #2.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>