Turn off LMEGANMONO when LMEGAN is turned off
Added code in READ_EMISSIONS_MENU (input_mod.f) to set
LMEGANMONO=.FALSE. when LMEGAN=.FALSE.  This is a bug fix.
If LMEGANMONO is turned on but LMEGAN is turned off, this will
cause an error (because arrays are not allocated, etc.).
A warning message will also be printed.

Also corrected spelling typos in files doc/Makefile and
co2_mod.f.  Made cosmetic changes in merra_a1_mod.f
and diag50_mod.f and added comments to define.h.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>