Minor fixes in megan_mod.f and gamap_mod.f
In routine INIT_MEGAN, we replaced the loop statement

     DO I = 15+BACK_ONE, 1, -1

with

     DO I = NUM_DAYS+BACK_ONE, 1, -1

because we read NUM_DAYS=0 instead of 15 days of data at the
start of a G-C simulation.  This prevents us reading in the
5 extra days of data.

Also in routine INIT_TRACERINFO (gamap_mod.f), only print out the
SOA tracers if LSOA=T.  This will prevent incomplete lines from
being written out to the tracerinfo.dat file.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>