New KPP scheme for SOA and standard chemistry
Mike Barkley (mpb14@leicester.ac.uk) wrote:
"Please note that with the new globchem.dat file (that fixes isoprene
nitrates and HNO4) that the ‘KPP/standard/’ programs must be rebuilt
using the procedures outlined in:
http://wiki.seas.harvard.edu/geos-chem/index.php/Interfacing_GEOS-Chem_with_KPP
and that it will not work with the‘out-of-the-box’beta code version:
v8-03-02. This is because the RIO2+NOàHNO3 reaction has been switched off
(‘D’). The globchem.dat works fine with the SMVGEAR solver."

Note that this is also true for the SOA simulation.


Signed-off-by: ccarouge <ccc@seas.harvard.edu>