Added ProTeX headers to several GEOS-Chem routines
Removed obsolete documentation headers in several
routines and and replaced them with the ProTeX headers,
which will facilitate auto-documentation.  Also updated the
list of GEOS-Chem files in the doc/Makefile_SrcDoc.mk directory.

Also saved the initial modifications to wetscav_mod.f for the
MERRA met fields.  We will return to this later.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>