Move code from WETDEP into several subroutines.
Split the code for:
- Rainout from the top level
- Rainout in the middle levels
- Washout in the middle levels
- No Downward precipitation
- Washout at level 1

into individual subroutines in wetscav_mod.f.  This is because
the modifications for the MERRA met fields require calling
the same sections of code more than once.  The easiest way
to deal with this is by introducing separate subroutines.

As of 9/17/10, the subroutines have been tested against a
reference code running with GEOS-5.  We get identical results.
The next step is to add in the new sequence for MERRA and do
additional cleanup.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>