Bug fix: Now can run starting from other times than 00 GMT
The MEGAN biogenic emissions reads in several days of A3 data
before the first timestep of a G-C simulation.  In routine
DO_OPEN_A3 (which determines if it is the right time to open
a new file), we have this IF statement:

      LOGICAL, SAVE :: FIRST    = .TRUE.

         ...

      ! Open A-3 file if it's 01:30 GMT,  or on the first call
      IF ( NHMS == 013000 .or. FIRST ) THEN
         DO_OPEN = .TRUE.
         ...

Therefore, when MEGAN was turned on, FIRST was set to .FALSE.
because we were reading more than one day of A3 data.  This
caused the IF statement above to fail if we wanted to start
a G-C simulation at any other time than 00 GMT.

The quick fix was to introduce an optional logical flag RESET,
such that:

      ! Reset the FIRST flag if
      IF ( PRESENT( RESET ) ) THEN
         IF ( RESET ) FIRST = .TRUE.
      ENDIF

This makes sure that the FIRST flag is not affected by the
prior A3 reads by the MEGAN emissions.

We now call OPEN_A3_FIELDS with RESET=.TRUE. from main.f for
the very first time we read in the A3 data (before the time-
stepping loop).

Also added ProTeX headrers to a3_read_mod.f.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>