Remove duplicate diagnostic categories in INIT_DIAGINFO
For whatever reason there were 2 sets of:

    CV-FLX-$
    TURBMC-$
    EW-FLX-$
    NS-FLX-$
    UP-FLX-$

defined in INIT_DIAGINFO (in gamap_mod.f).  One set of these
categories had an offset of 3000 (i.e. SPACING*3) and another had
an offset of 4000 (i.e. SPACING*4).

Because these are all in kg/s, the correct offset should be
3000.  The duplicate definitions were commented out (to be
removed shortly).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>