Now can run MERRA sims from start times than 00 GMT
The MEGAN biogenic emissions reads in several days of
MERRA A1 data before the first timestep of a G-C simulation.
In routine DO_OPEN_A1 (which determines if it is the right
time to open a new file), we have this IF statement:

      LOGICAL, SAVE :: FIRST    = .TRUE.

         ...

      ! Open A1 file if it's 00:30 GMT,  or on the first call
      IF ( NHMS == 003000 .or. FIRST ) THEN
         DO_OPEN = .TRUE.
         ...

Therefore, when MEGAN was turned on, FIRST was set to .FALSE.
because we were reading more than one day of A1 data.  This
caused the IF statement above to fail if we wanted to start
a G-C simulation at any other time than 00 GMT.

The quick fix was to introduce an optional logical flag RESET,
such that:

      ! Reset the FIRST flag if
      IF ( PRESENT( RESET ) ) THEN
         IF ( RESET ) FIRST = .TRUE.
      ENDIF

This makes sure that the FIRST flag is not affected by the
prior A1 reads by the MEGAN emissions.

We now call OPEN_A1_FIELDS with RESET=.TRUE. from main.f for
the very first time we read in the A1 data (before the time-
stepping loop).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>