Wet deposition modifications for MERRA
Add Helen Amos' revisions for wetscav_mod.f, including:

(1) Update Q and QDOWN to be consistent with MERRA met

(2) Change WASHFRAC_LIQ_GAS to be a subroutine

(3) Disable new definition for ALPHA for partial evaporation
    within routine DO_WASHOUT_ONLY

Also minor modifications in gc_type_mod.f and convection_mod.f.
At this stage, the MERRA convection routine is not yet complete,
and is awaiting the addition of code for re-evaporation.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>