Prevent NaN's when computing L2S and L3S in CHEM_SO2
When computing L2S and L3S, we need to extend the IF condition
from:

   IF ( IS_SAFE_EXP( XX ) ) THEN

to:

   IF ( IS_SAFE_EXP( XX ) .and. ABS( XX ) > 0d0 ) THEN

When computing L2S, the term XX will be zero if SO2_ss == H2O20.
This can happen if SO2_ss and H2O20 are both equal to zero.
Therefore, we need to skip computation of L2S in the IF block
and shunt to the ELSE block, where L2S will be assigned a
default value.

Ditto for the computation of L3S, which can also incur a
div-by-zero error if SO2_ss == O3.

Jamin Koo discovered this error.

Signed-off-by: Jamin Koo <jaminkoo@mit.edu>
Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>