Save PEDGE-$ in all timeseries diagnostics
Due to historical baggage, the timeseries diagnostics only saved
out Psurface - PTOP for the first level.

This is now changed this so that we now save out PEDGE-$
(i.e. the pressure at the bottom edge) of each requested level.

Also modified "gamap_mod.f" to write out the proper information
for the surface pressure to the "tracerinfo.dat" file if any
of the timeseries diagnostics are selected.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>