Apply Helen Amos' fix for washout @ surface
In the routine WETDEP_MERRA, there was a typo where FTOP was
set to zero at the end of a DO loop over vertical levels.
This caused washout to not be applied at the surface level.

Fix this by setting FTOP = F_RAINOUT + F_WASHOUT within the
DO loop.

Also added the ProTeX-style headers to findmon.f.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>