Fixed errors in time_mod.f for MERRA met data
The routines SET_BEGIN_TIME and SET_END_TIME in time_mod.f
flagged an error if the start and end dates were prior to
1985/01/01.  However, the MERRA met data archive goes back
in time prior to 1985.

Therefore, adjust the test for THISNYMDb in these two routines
so as not to trip an error if the date is prior to 1985.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>