Fixes for starting GEOS-Chem from arbitrary times
Added fix to GET_TIME_AHEAD (time_mod.f) in case the HHMMSS
value is returned as 240000.  In that case, set it to 000000
and then call JULDAY/CALDATE to find out the date 1 day ahead.
This condition can be a side-effect of roundoff error.

In main.f, now call OPEN_MERRA_I6_FIELDS at the start of the
simulation with the date returned from GET_FIRST_I6_TIME.

Fixed comments in merra_i6_mod.f and julday_mod.f

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>