Changed retention efficiency for POPs for rainout
I realized I had changed the retention efficiency (from 0.02 to 0 for
temps < 268 K) for washout only. Now I've changed it for rainout as well.
(POPs are not very water-soluble. Expect them to be excluded from cold water
more efficiently than some of the smaller compounds the 0.02 number was
oroginally based on).