TOMAS: Synch updates from GeosCore with code in GeosTomas
The code in GeosTomas was synched up with recent updates from
the code in GeosCore.  This is to ensure that the TOMAS code
is consistent with the standard GEOS-Chem code in advance of
the v9-01-01 release.


NOTE TO TOMAS DEVELOPERS:
-------------------------
The wet scavenging algorithms in wetscav_mod.f have
been totally rewritten.  We therefore now have the
following modules:

  wetscav_mod.f_orig : The "original", untouched code
                        from v8-03-02
  wetscav_mod.f_new  : The "new" wet scavenging code
                        from v9-01-01
  wetscav_mod.f      : Right now this is a copy of
                        wetscav_f.orig

You will therefore have to put all of the TOMAS microphysics
into the wetscav_mod.f_new file in routines WETDEP and
WETDEP_MERRA.  This task is beyond the scope of the GEOS-Chem
Support Team.

  -- Bob Yantosca (21 Jan 2011)

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>