Patch to avoid double-counting biofuel emissions over Asia
Claire Carouge found a bug in biofuel_mod.f about Streets
emissions. Initially we used Streets 2000 for years <= 2005,
then Streets 2006 for year >= 2006. But this was changed
because of discrepancies in the total emissions between 2005
and 2006. So now we use Streets 2000 for years <=2000 and
Streets 2006 for years >= 2001. The problem is that Streets
2006 contains biofuels and not Streets 2000. So in
biofuel_mod.f,  we must change the following IF condition to
( SIM_YEAR >= 2001 ) to avoid double counting biofuel emissions
over Asia.

Signed-off-by: Claire Carouge <ccc@seas.harvard.edu>
Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>