Now point to updated CDIAC CO2 emissions
Robert Andres has released updated annual and monthly CO2
emissions from the CDIAC inventory.  These are stored in the

  GEOS_2x2.5/CO2_201003/fossilfuel_andres/annual_v2010
  GEOS_2x2.5/CO2_201003/fossilfuel_andres/monthly_v2010

  GEOS_4x5/CO2_201003/fossilfuel_andres/annual_v2010
  GEOS_4x5/CO2_201003/fossilfuel_andres/monthly_v2010

In order to not overwrite the existing annual emissions data,
(for the sake of backwards compatibility), we have uploaded
these new directories to the GEOS-Chem data directory structure
and have modified co2_mod.f so that the code will now read from
the annual_v2010/ and monthly_v2010/ directories.

Thanks to Ray Nassar for sending us the data.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>