Now use "D" exponents for constants in vdiff_mod.f90
Now explicitly force use of the "D" exponents (i.e. 0.d0
instead of 0.) in many locations.  This will prevent errors
from occurring on the PGI compiler.

This fix was first made in the APM branch.  We are including
this in v9-01-01-Patch-VDIFF patch so that G-C users can
update their codes with this ASAP.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>