Update to Holmes et al. 2010 version
Several changes to mercury code to match what was included in Holmes et al.
(2010). Most of these have to do with consistency of land/water/ice
characterization. Also included are changes to Hg/CO ratio for biomass
burning emissions and deposition of only 60% of Hg to snowpack that can
re-emit.

Signed-off-by: Jenny Fisher <jafisher@fas.harvard.edu>