Bug fix in ND54 "time-in-tropopause" diagnostic
The ND54 diagnostic (time-in= ND54 is updated in DIAG1, ]'
which is called every chemistry timestep (e.g. 60 min).

However, when we save out the data to disk in diag3.f,
we divide by the wrong scale factor:

     SCALEDYN   = DBLE( GET_CT_DYN()  ) + 1d-32
     SCALEDIAG  = DBLE( GET_CT_DIAG() ) + 1d-32
     ...

         DO L = 1, LD54
            ARRAY(:,:,L) = AD54(:,:,L) / SCALEDYN
         ENDDO

This results in the ND54 diagnostic being precisely twice
as small as it should be.  The maximum value is 0.5 instead
of 1.

The fix is to simply to divide by SCALEDIAG (which is updated
each chemistry timestep) instead of SCALEDYN.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>