Fix to variable tropopause
Fix to variable tropopause by Daven Henze.  Following a suggestion
from Dylan Jones, when a grid cell that appears anew in the troposphere,
we search downward in the column until we find a cell that was in the
troposphere the previous time step, and use information from that cell
for partitioning.

Signed-off-by: Matthew Cooper <cooperm2@dal.ca>