Now use correct indexing when archving into CSAVE array
Corrected a typo in in partition.f.  This line of code:

   CSAVE(JLOOP,N) = CSPEC(JLOOP,CSAVEID_JJ(N))

needed to be changed to:

   CSAVE(JLOOP,N) = CSPEC(JLOOP_P,CSAVEID_JJ(N))

where JLOOP_P is the 1-D index for the (I,J,L) at the last
chemistry timestep.  If this grid box was in the stratosphere,
then JLOOP_P will be assigned to (I,J,LTEST), which is the
highest grid box in the column (I,J) that is still tropospheric.

Therefore, CSAVE now archives only tropopsheric concentrations,
which result in proper partitioning.

This typo caused the fix for the dynamic tropopause not to
be applied.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>