Corrected mis-indexing error in restart.cspec.YYYYMMDDhh
Jingqiu Mao found a problem with the

  The index of species in CSPEC is changed by turning on P/L diagnostics (ND65).

    Without P/L diagnostics, CSPEC is indexed as
     ( *, active species + inactive species )

    With P/L diagnostics, CSPEC is indexed as
     ( *, active species + P/L diagnostics + inactive species )

  So if you read in a restart.cspec.YYYYMMDDhh file and switch
  on P/L diagnostics, model will set lots of inactive species
  to zero, such as CH4, N2 etc. Ozone and H2O are not affected
  as they are reinitialized in the code.  But this is a big
  problem if you have zero CH4 in the code.

The fix for this problem is to skip the P/L diagnostic families,
which are denoted by the IFAM array.  Therefore, we now test
whether the current CSPEC species number belongs to IFAM before
writing each species to the restart file.  If so, then we skip
writing that species to the restart.cspec.YYYYMMDDhh file.

A similar check is also made when reading the species from the
restart file at the start of a simulation.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>