Prevent out-of-bounds error in biomass_mod.f
In biomass_mod.f, the IDBCO flag was used as an array index.
But for some offline simulations, IDBCO=0, which can cause an
out-of-bounds errors.  We fixed this by making sure IDBCO > 0
before using it to index an array.  We also took advantage of
this opportunity to add the ProTeX headers to biomass_mod.f
and updated the doc/Makefile_SrcDoc.mk makefile accordingly.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>