Correct the 'tracer 12 is not an emitted species' message
In GeosCore/tracerid_mod.f, we had this line of code:

   IF ( IDTRCHO > 0 ) THEN
      NEMANTHRO = NEMANTHRO + 1
      IDERCHO  = NEMANTHRO
   ENDIF

This caused GEOS-Chem to think that RCHO is an emitted tracer.
However, the RETRO emissions do not use RCHO, and thus we got
a spurious warning message.

Commenting out this IF block prevents this warning message from
being printed.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>