Replaced #include statements with USE statements in GeosTomas
Because we have replaced all of the *.h and CMN* files in
Headers with F90 modules, we need to update the files in
GeosTomas accordingly.  References such as #include "CMN_SIZE"
have now been replaced with "USE CMN_SIZE_MOD", etc.

Also made some cosmetic changes (mostly to line up comments in
the same column) to files in the GeosCore directory.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>