Removed obsolete commented out code
In many G-C routines, there are sections of code labeled
"Prior to MM/DD/YY".  These indicate the original code that
was later replaced by an update.  In order to preserve
the clarity of the code, we have removed these comments
from several routines in the GeosCore, GeosApm, and GeosUtil
directories.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>