Bug fixes for APM
Gan Luo discovered that GEOS_5 was undefined in module WETSCAV_MOD for the
v9-01-02n APM simulation. This was fixed by adding:

  include "define.h"

in wetscav_mod.F and apm_wets_mod.F.

Also, to allow APM to run with GFED3 the dependencies to make the GFED3
code were added to GeosApm/Makefile.

Signed-off-by: Melissa Payer <mpayer@rhea.as.harvard.edu>