Centralize the chemistry time step
The current setup for 2x2.5 is:

  timestep 1 : transport (15min) + Chemistry (60min)
  timestep 2: transport (15 min)
  timestep 3 : transport (15min)
  timestep 4: transport (15 min)--------------------------->output
  timestep 5 : transport (15min) + Chemistry (60min)
  timestep 6: transport (15 min)
  timestep 7 : transport (15min)
  timestep 8: transport (15 min)--------------------------->output

Lin Zhang provided fixes in GeosUtil/time_mod.F and GeosCore/diag_pl_mod.F
so that the chemistry time step is in the center of transport time steps
(i.e., TCTTT becomes TTCTT).

Bob Y. also provided a fix for calling the planeflight diagnostic in chemdr.F.
Moving the chemistry time step to the middle of the transport time steps
caused the model to crash when it calls SETUP_PLANEFLIGHT. The fix provided
by Bob calls SETUP_PLANEFLIGHT on the first chemical time step of a new day
by testing to see when the date is not the same as the previous date.

Signed-off-by: Melissa Payer <mpayer@rhea.as.harvard.edu>