Add SUNCOS30 array for cos(SZA) at 30min after each GMT hour
We recently realized that the sun angles are always computed
at 30 minutes after the current time, and not at 30 minutes
after the hour.

For example, in the v9-01-02q fix, the current scenario is:

              00:00  00:30  GMT
4x5           T(PC)  T
                     *

              00:00  00:15  00:30  00:45  GMT
2 x 2.5       T      T(PC)  T      T
                                   *

              00:00  00:10  00:20  00:30  00:40  00:50  GMT
0.5 x 0.666   T      T      T(PC)  T      T      T
                                                 *

This is because we add an offset to the local time in routine
COSSZA (in dao_mod.F)

      ! 1/2 of the time interval (normally the chemistry timestep)
      ! for computing SUNCOS.  Convert from minutes to hours.
      OFFSET = GET_TS_SUN_2() / 60d0

         ...

            ! Local time at box (I,J) [hours]
            TIMLOC = GET_LOCALTIME( I, OFFSET )

This is just offsetting the local time but not fixing it to a
SPECIFIC time of day.

Therefore, to make the previous fixes to v9-01-02q work, we
need to use a cos(SZA) computed consistently at 30m after each
GMT hour.  This is now contained in the SUNCOS30 array, which
is returned by routine COSSZA (in dao_mod.f).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>