SUNCOS_MID now returns cos(SZA) @ midpt of chem timestep
Rename SUNCOS30 to SUNCOS_MID.  SUNCOS_MID now holds the
values of the cos(SZA) computed at the midpoint of the
chemistry timestep (not always at 00:30 as before).  This is
now computed consistently depending for any combination
of transport & chemistry timesteps.

In time_mod.f, we also have modified ITS_TIME_FOR_EMIS so that
it is similar to ITS_TIME_FOR_CHEM.  This will ensure that
emissions are done on the same timestep when chemistry is
done.

Furthermore, we have now added error checks in input_mod.f
to make sure that the inputted emission timestep is the same
as the inputted chemistry timestep.  Otherwise, we could have
a situation where emissions & drydep occur on one timestep but
chemistry occurs on a different timestep.

We have also added a similar error check in input_mod.f to
ensure that the convection timestep is the same as the
dynamic timestep.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>