Modify ITS_A_NEW_MONTH for central chemistry time step
Rewrite the logic of function ITS_A_NEW_MONTH (in time_mod.F)
so that it computes the first time to read data in a new month
that is consistent w/ the central chemistry time step option.

The default behavior will return TRUE at the midpoint of the
first chemistry timestep (= the first emissions timestep)
of the month.  For simulations starting at an arbitrary
midmonth date & time, this will return TRUE at the first
chemistry timestep after the start of the simulation.

One can revert to the previous behavior (returning TRUE at
00:00 GMT on the first day of the month, or at the start time
of the simulation) with this calling sequence:

   ITS_A_NEW_MONTH( NO_CCTS=.TRUE. )

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>