Set small K_OX and K_RED to 0
With MERRA, occasional values in met fields were causing K_OX to be very
small (10^-14), causing the run to stop in the redox reaction computation
when the combined rate constant quantity was below the threshold value.
Now we set very small K_OX and K_RED values to 0 to prevent this problem.


Signed-off-by: Jenny Fisher <jafisher@fas.harvard.edu>
               Chris Holmes <cdholmes@uci.edu>