Now add check for netCDF installation integrity
Add the TestNcdfUtil.F90 program (which is the same
program from the NcdfUtilities package, just renamed
for consistency) to the NcdfUtil directory.

Also add the ncdfcheck Makefile target to ./Makefile
./GeosCore/Makefile.  This will do the following calls:

  make -C $(NCDF) lib
  make -C $(NCDF) check

which will build the NcdfUtil package and run a test
program to check the integrity of the netCDF library
installation.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>