Now echo info after reading GEOS-5.7.x met data
Added several WRITE statements in the various routines
that read GEOS-5.7.x met data from disk.  This will
denote the number of met fields that was read from disk
at a given date and time.

Also added an extra error check to prevent re-reading
GEOS-5.7.x A1 and A3* met data for the same date and,
in the same way as is done for MERRA met fields.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>