Add error trap for PARANOX LUT of fracnox
Previously, the error trap in INTERPOLATE_LUT for fracnox < 0 only printed
message to log file, but run would continue along. A call to ERROR_STOP has
been added so that if there is a bad value of fracnox, the run will stop
with an error message.

Geert Vinken wrote:

  This was an error trap I build in, to flag when I encountered possible
  combinations not present in my LUT [look up table]. I think all possible
  combinations should be included, so I suggest stopping with an error if
  this in encountered. Perhaps you can extend the trap to stop in fracnox
  is .lt. 0 or .gt. 1 (this would also catch errors with corrupt LUTs).

Cosmetic changes were also made to paranox_mod.F to improve readability
of the code.

Signed-off-by: Melissa Payer <mpayer@seas.harvard.edu>