Synched GeosTomas code with GeosCore code, again
Made several modifications to files in GeosTomas so as
to synchronize them with the files in GeosCore.  In
particular, modified the drydep_mod.F, seasalt_mod.F,
sulfate_mod.F and Makefile in GeosTomas.  Added ProTeX
headers to be consistent w/ files in GeosCore.  Also
corrected several typos & errors from last time.

NOTE: The file GeosTomas/wetscav_mod.F is the current
version from GeosCore.  This was updated recently and will
have to be modified for TOMAS.  This is beyond the
capability of the GC Support Team.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>