Remove duplicate allocation of CLDTOPS
When the allocation of dao_mod.F module arrays was
reorganized, CLDTOPS was allocated twice, once in
INIT_DAO_GEOS5 and again in INIT_DAO_DERIVED.  This
caused an allocation error.

CLDTOPS only needs to be allocated from routine
INIT_DAO_DERIVED, so we can remove the extraneous
allocation in INIT_DAO_GEOS5.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>