Seek to avoid roundoff errors in olson_landmap_mod.F90
In COMPUTE_OLSON_LANDMAP, now use REAL*4 for computations
to avoid some roundoff errors at 0.5 x 0.666 grids.
Bob Yantosca validated this update with 4 x 5 and
2 x 2.5 simulations.

NOTE: for the 0.5 x 0.666 grid, you may see differences
of 1/1000 in the IUSE and FRCLND arrays.  We are still
investigating why this occurs, but it seems like it is
due to roundoff error caused by the longitude spacing
being an irrational number (0.66666666666....).

Add a note to the comments denoting that 0.5 x 0.666
is still under investigation.  Differences in 1/1000
in IUSE and FRCLND shouldn't change the results of
the simulation by very much.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>