Remove Fg and Fp PL-HG2-$ diagnostics, and HgP IJ-AVG-$ diagnostic.
Fg (fraction of Hg2 in gas phase) and Fp (graction Hg2 to in
particle phase) were redundant. Fg and Fp can  be calculated from
from RGM and PBM.  The IJ-AVG-$ diagnostic for HgP was
removed because HgP is now emitted as Hg2, so it's an empty
diagnostic (i.e. all zeros).

Signed-off-by: Helen Amos <hamos@rhea.as.harvard.edu>