Bug fix: don't reset LMODISLAI or LAVHRLAI when LEMIS=F
Bob Yantosca introduced a minor bug such that both the
LMODISLAI and LAVHRRLAI switches were turned off when
LEMIS was turned off (i.e. when shutting off emissions).
Because LAI is used in other areas of the code, we cannot
assume this.

We have now commented out the offending lines of code.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>