Modifications for using MODIS LAI at 0.25 x 0.25
Rewrote program logic in routine READ_MODIS_LAI (in
modis_lai_mod.F90) to use the first or last years
of data if the requested year lies outside of the
MODIS data range.  This is done when we are using MODIS
at 0.25 x 0.25 inputs (to match up w/ the Olson 2001
land map).

Also made a minor change in mapping_mod.F90 to try to
prevent out-of-bounds errors when computing the
FINE_PER_COARSE quantity (which is used to dimension
the mapping object).  Mapping_mod.F90 now relies on
logical_mod.F90 (it references the USE_OLSON_2001
parameter).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>