Bug fix in ARSL1K
In the tropospheric bromine chemistry mechanism, a sticking coefficient (STKCF)
of near zero is passed to the function ARSL1K for non-sulfate, non-sea salt
aerosol. The IF statement modified in GEOS-Chem v8-02-04 to include checks for
zero values of DENAIR, SQM, STK, and STKCF resulted in the reaction rate being
set to the default value of 1.0d-3. A 1-month benchmark for July 2005 indicated
that the simulated BrO was a little more than twice the expected zonal mean.
Modifying the default value from 1.0d-3 to 1.0d-30 resulted in reasonable
simulated BrO values.

Mat Evans found that changing the default value to 1.0d-30 did not produce a
large change in the standard 43-tracer full-chemistry simulation.

Mat wrote:

    I've re-run two 2 month simulation. One with the error handling value of
    1e-3 (standard) and one with it being 1e-30. There are 5127 time and
    space points where the model traps the problem and invokes the 1e-3 or
    1e-30 value. There are 30*24*2*37*72*46 (roughly 200 million) time and
    space points when the error could have occurred so we are looking at a
    relatively infrequent event.

    The simulations show virtually no difference between the two simulations.

    mean and stddev ratio of all grid boxes with and without the fix are shown
    below
    NOx     0.999996  0.000409291
    Ox      1.00000   1.27233e-05
    O3      1.00000   1.52284e-05
    PAN     0.997849  0.0111997
    CO      1.00000   4.21768e-06
    ALK4    0.990514  0.0351941
    ISOP    0.999979  0.0108033
    H2O2    0.992067  0.0264659
    DST1    1.00000   0.00000
    HO2     0.999996  0.00309464
    OH      1.00003   0.00767954

    So although there are some differences they are very minor. For
    completeness we should put this in as a bug fix (make the error value
    1d-30 rather than 1d-3). But it is not a major problem.

Signed-off-by: Melissa Payer <mpayer@seas.harvard.edu>