Second pass in connecting met fields to GC_MET_LOCAL
The 2D met field arrays are connected to GC_LOCAL_MET derived type using
a top-down method. Note that this still needs to be cleaned up. The entire
type LOCAL_MET may not need to be passed from top down, depending on how
many of the met fields are utilized in a given subroutine.

Signed-off-by: Melissa Payer <mpayer@seas.harvard.edu>